Feature/extract input listener

+++ src/main/java/fr/snapgames/bgf/UIText.java
@@ -46,7 +46,7 @@ public class UIText extends GameObject {
 	public void render(Graphics2D g) {
 		g.setFont(font);
 
-		Render.drawOutlinedString(g, x, y, text, thickness, Color.WHITE, Color.BLACK);
+		Render.drawOutlinedString(g, (int)x, (int)y, text, thickness, Color.WHITE, Color.BLACK);
 	}
 
 	public UIText setThickness(int thickness) {
